+++ app/javascript/controllers/flash_controller.js
@@ -2,7 +2,7 @@ import { Controller } from "@hotwired/stimulus"
 
 export default class extends Controller {
   static values = {
-    duration: { type: Number, default: 25000 }
+    duration: { type: Number, default: 5000 }
   }
 
   connect() {